+++ pyzoo/zoo/pipeline/api/net/tf_optimizer.py
@@ -390,6 +390,10 @@ class TFOptimizer:
         TFOptimizer is used for distributed training of TensorFlow
         on Spark/BigDL.
 
+        Note that if grads and variables are not None, then they need to be sorted by name
+        if you want to use multiple optimization methods for a TensorFlow model according to
+        variable names.
+
         :param loss: The loss tensor of the TensorFlow model, should be a scalar
         :param optim_method: the optimization method to be used, such as bigdl.optim.optimizer.Adam
         :param sess: the current TensorFlow Session, if you want to used a pre-trained model, you
