--- zoo/src/test/resources/mnist_keras/graph_meta.json
@@ -1,1 +1,0 @@
-{"output_names": ["gradients/dense/BiasAdd_grad/BiasAddGrad:0", "gradients/dense/MatMul_grad/MatMul_1:0", "gradients/dense_1/BiasAdd_grad/BiasAddGrad:0", "gradients/dense_1/MatMul_grad/MatMul_1:0", "gradients/dense_2/BiasAdd_grad/BiasAddGrad:0", "gradients/dense_2/MatMul_grad/MatMul_1:0", "dense_2/Softmax:0", "ToFloat:0", "dense_2/Softmax:0", "ToFloat_1:0", "ToFloat_2:0", "loss/mul:0"], "input_names": ["input_1:0", "dense_2_target:0"]}
